templates/cloud-run/python/requirements.txt
fastapi
uvicorn[standard]
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-exporter-gcp-trace
opentelemetry-instrumentation-fastapi
google-cloud-trace
google-cloud-logging
